add cms route in app initializer
